Moment matrix generation

Functions

QuantumNPA.npa_moment — Method

npa_moment(operators)

Construct the NPA moment matrix, given a vector of operators (monomials or polynomials). The moment matrix returned is a representation of the real part of the moment matrix, i.e., the moment matrix plus its complex conjugate divided by two. It is returned in the form of a polynomial with sparse matrices as the monomials.

Examples

julia> gamma = npa_moment([Id, A1, A2, B1, B2]);

julia> gamma[A1]
5×5 SparseArrays.SparseMatrixCSC{Float64, Int64} with 2 stored entries:
  ⋅   1.0   ⋅    ⋅    ⋅
 1.0   ⋅    ⋅    ⋅    ⋅
  ⋅    ⋅    ⋅    ⋅    ⋅
  ⋅    ⋅    ⋅    ⋅    ⋅
  ⋅    ⋅    ⋅    ⋅    ⋅

julia> QuantumNPA.repack(gamma)
5×5 Matrix{Polynomial}:
 Id  A1     A2     B1     B2
 A1  Id     A1 A2  A1 B1  A1 B2
 A2  A1 A2  Id     A2 B1  A2 B2
 B1  A1 B1  A2 B1  Id     B1 B2
 B2  A1 B2  A2 B2  B1 B2  Id

julia> gamma = npa_moment([Id, A1 + A2, A1 - A2]);

julia> gamma[A1]
3×3 SparseArrays.SparseMatrixCSC{Float64, Int64} with 4 stored entries:
  ⋅   1.0  1.0
 1.0   ⋅    ⋅
 1.0   ⋅    ⋅

julia> gamma[A2]
3×3 SparseArrays.SparseMatrixCSC{Float64, Int64} with 4 stored entries:
   ⋅   1.0  -1.0
  1.0   ⋅     ⋅
 -1.0   ⋅     ⋅

julia> QuantumNPA.repack(gamma)
3×3 Matrix{Polynomial}:
 Id       A1 + A2             A1 - A2
 A1 + A2  2.0 Id + 2.0 A1 A2  0
 A1 - A2  0                   2.0 Id - 2.0 A1 A2

QuantumNPA.npa_moment — Method

npa_moment(source, level)

Constructs the NPA moment matrix at the given level of the hierarchy, taking all degree 1 monomials appearing in source as the level 1 operators. source can be a monomial, polynomial, or collection containing monomials, polynomials, or sub-collections. The level can be a nonnegative integer or a string, such as "1 + A B".

QuantumNPA.npa2sdp — Method

npa2sdp(expr, level; eq=[], ge=[])

Generate the NPA relaxation for a given quantum optimisation problem, at the indicated level of the NPA hierarchy.

This basically uses the npa_moment function to generate the moment matrix and then calls npa2sdp again with the moment matrix as the second argument, which takes the real part of the problem and then eliminates the equality constraints by substitution.

Example

julia> S = A1*(B1 + B2) + A2*(B1 - B2)
A1 B1 + A1 B2 + A2 B1 - A2 B2

julia> (objective, moments) = npa2sdp(S, 1, eq=[A2*B1 - Id]);

julia> objective
Id + A1 B1 + A1 B2 - A2 B2

julia> moments[1][Id]
5×5 SparseArrays.SparseMatrixCSC{Float64, Int64} with 7 stored entries:
 1.0   ⋅    ⋅    ⋅    ⋅
  ⋅   1.0   ⋅    ⋅    ⋅
  ⋅    ⋅   1.0  1.0   ⋅
  ⋅    ⋅   1.0  1.0   ⋅
  ⋅    ⋅    ⋅    ⋅   1.0

julia> QuantumNPA.repack(moments[1])
5×5 Matrix{Polynomial}:
 Id  A1     A2     B1     B2
 A1  Id     A1 A2  A1 B1  A1 B2
 A2  A1 A2  Id     Id     A2 B2
 B1  A1 B1  Id     Id     B1 B2
 B2  A1 B2  A2 B2  B1 B2  Id